backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.2
google-cloud-aiplatform==1.71.1
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
